+++ src/index.js
@@ -1,5 +1,6 @@
 /* global vegaTooltip, vegaEmbed, vega, vl */
 import '@babel/polyfill';
+const pLimit = require('p-limit');
 
 
 // stackoverflow
@@ -20,12 +21,9 @@ function filterOutliers(someArray) {
 }
 
 // create spec
-async function process(data) {
-    const requests = await Promise.all(data);
-    const ret = await Promise.all(requests.map(m => m.json()));
-    let d = [].concat(...(ret.map(m => m.builds)));
+async function process(d) {
     d = d.map(m => ({
-        message: m.commit.message,
+        message: (m.commit || {}).message,
         branch: (m.branch || {}).name,
         duration: m.duration,
         number: m.number,
@@ -68,6 +66,36 @@ async function process(data) {
 function timer(ms) {
     return new Promise(r => setTimeout(r, ms));
 }
+function throttleActions(listOfCallableActions, limit) {
+  // We'll need to store which is the next promise in the list.
+  let i = 0;
+  let resultArray = new Array(listOfCallableActions.length);
+
+  // Now define what happens when any of the actions completes. Javascript is
+  // (mostly) single-threaded, so only one completion handler will call at a
+  // given time. Because we return doNextAction, the Promise chain continues as
+  // long as there's an action left in the list.
+  function doNextAction() {
+    if (i < listOfCallableActions.length) {
+      // Save the current value of i, so we can put the result in the right place
+      let actionIndex = i++;
+      let nextAction = listOfCallableActions[actionIndex];
+      return Promise.resolve(nextAction())
+          .then(result => {  // Save results to the correct array index.
+             resultArray[actionIndex] = result;
+             return;
+          }).then(doNextAction);
+    }
+  }
+
+  // Now start up the original <limit> number of promises.
+  // i advances in calls to doNextAction.
+  let listOfPromises = [];
+  while (i < limit && i < listOfCallableActions.length) {
+    listOfPromises.push(doNextAction());
+  }
+  return Promise.all(listOfPromises).then(() => resultArray);
+}
 
 async function graph(e) {
     if (e) {
@@ -78,37 +106,49 @@ async function graph(e) {
     const end = document.getElementById('end').value;
     document.getElementById('view').innerHTML = 'Loading...';
 
-    window.history.replaceState({}, '', `?repo=${repo}&start=${start}&end=${end}`);
 
 
-    const data = [];
-    const limit = 100;
+    const nBuilds = 100;
     const headers = new Headers({ 'Travis-API-Version': '3' });
-    const res = await fetch(`https://api.travis-ci.org/repo/${repo}/builds?sort_by=id:desc`, { headers });
+    const prefix = `https://api.travis-ci.org/repo/${repo}/builds`;
+    const res = await fetch(`${prefix}?sort_by=id:desc`, { headers });
     if (res.status !== 200) {
         throw `Error ${res.status}: ${res.statusText}`;
     }
+    window.history.replaceState({}, '', `?repo=${repo}&start=${start}&end=${end}`);
     const resjs = await res.json();
     document.querySelector('#totalBuilds').innerHTML = `Total number of builds: ${resjs.builds[0].number}`;
-    for (let i = +start; i <= +end; i += limit) {
-        const builds = fetch(`https://api.travis-ci.org/repo/${repo}/builds?limit=${limit}&offset=${i}&sort_by=id`, { headers });
-        data.push(builds);
-        document.getElementById('view').innerHTML = `Loading build ${i}...`;
-        await timer(500);
+ 
+    let input = [];
+    const limit = pLimit(3);
+
+    for (let i = +start; i <= +end; i += nBuilds) {
+        input.push(limit(() => {
+            document.getElementById('view').innerHTML = `Loading build ${i}...`;
+			return fetch(`${prefix}?limit=${nBuilds}&offset=${i}&sort_by=id`, { headers })
+        }));
     }
-    return process(data);
+	 
+	const result = await Promise.all(input);
+    const ret = await Promise.all(result.map(m => m.json()));
+    const builds = ret.map(m => m.builds);
+    return process([].concat(...builds));
 }
 
-document.querySelector('form').addEventListener('submit', graph);
+function catchgraph(e) {
+    graph(e).catch(error => {
+        document.querySelector('#view').innerHTML = `<div class="alert alert-warning">${error}</div>`;
+    });
+}
 
+document.querySelector('form').addEventListener('submit', catchgraph);
 const params = new URLSearchParams(window.location.search.slice(1));
 if (params.get('repo')) document.getElementById('repo').value = params.get('repo');
 if (params.get('start')) document.getElementById('start').value = params.get('start');
 if (params.get('end')) document.getElementById('end').value = params.get('end');
 
-graph().catch(error => {
-    document.querySelector('#view').innerHTML = `<div class="alert alert-warning">${error}</div>`;
-});
+
+catchgraph();
 
 
 document.getElementById('versions').innerHTML += `vega: ${vega.version} vega-lite: ${vl.version}`;